+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Change KDE Plasma from version 5 to 6.
 - Change the initial 'yay' installation from version 12.1.3 to 12.3.5.
 - Change GRUB installation to not modify UEFI variables on the motherboard.
+- Change GRUB to use partition UUIDs instead of Linux UUIDs.
 
 ## 4.0.0 - The Conversion Update